fixed #24 treeViewWindow
