+++ src/main.py
@@ -33,7 +33,7 @@ __credits__ = ['Terry Moore', 'Velmurugan Selvaraj']
 __date__ = '2021/07/12'
 __maintainer__ = ['Seenivasan V', 'Vinay N']
 __status__ = 'Production'
-__version__ = '2.3.14'
+__version__ = '2.4.0'
 ##############################################################################
 # Utilities
 ##############################################################################
